pyproject.toml
[project]
name = "natshell"
version = "0.1.13"
description = "Natural language shell interface — a local-first agentic TUI powered by a bundled LLM"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Nicholas B Considine" }
]
keywords = ["shell", "natural-language", "tui", "llm", "agent", "cli", "linux", "macos"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
"Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Shells",
    "Topic :: Utilities",
]

dependencies = [
    "textual>=1.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
natshell = "natshell.__main__:main"

[project.optional-dependencies]
local = ["llama-cpp-python>=0.3.0", "huggingface-hub>=0.24"]
mcp = ["mcp>=1.0.0"]
dev = [
    "llama-cpp-python>=0.3.0",
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[project.urls]
Repository = "https://github.com/Barent/natshell"
Issues = "https://github.com/Barent/natshell/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
natshell = ["config.default.toml", "ui/styles.tcss"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"src/natshell/agent/system_prompt.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
